fix(AbstractSnapshot.php): fix a PHPUnit v5 incompatibility issue
Previous versions of PHPUnit do not implement the `dataName` method and would return an empty string
from the `dataDescription` method if the data name is empty. In this empty case would fall a
non-named data set with an index of `0`. This fix used Codeception built-in ReflectionHelper class
to read the private property starting from the test case and "climbing" the inheritance tree.